reorganize the auth components
