+++ web/src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 import { HomeComponent } from './component/home/home.component';
-import { RegisterComponent } from './component/register/register.component';
-import { LoginComponent } from './component/login/login.component';
+import { RegisterComponent } from './component/auth/register/register.component';
+import { LoginComponent } from './component/auth/login/login.component';
 
 export const routes: Routes = [
   {
