+++ web/src/app/component/auth/register/register.component.scss
